// src/api/openapi/model/scopeEnum.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const ScopeEnum = {
    Admin: 'admin',
    Office: 'office',
    Workshop: 'workshop'
} as const;
export type ScopeEnum = typeof ScopeEnum[keyof typeof ScopeEnum];