// src/api/openapi/models/userCreate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserCreate { 
    chat: boolean;
    cost: number;
    dial: string;
    email: string;
    employee: boolean;
    firstname: string;
    handy: string;
    hours: boolean;
    innovaphone_pass: string;
    innovaphone_user: string;
    notifications: boolean;
    office: boolean;
    password: string;
    position: string;
    secondname: string;
    tel: string;
}